refactor(engine): Move Tests to HistoryCleanupTest
* Code Review point

Related to: #2728
