+++ engine/src/test/java/org/camunda/bpm/engine/test/bpmn/async/JobRetryCmdWithDefaultPropertyTest.java
@@ -16,24 +16,19 @@
  */
 package org.camunda.bpm.engine.test.bpmn.async;
 
-import static org.camunda.bpm.model.xml.test.assertions.ModelAssertions.assertThat;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.fail;
 
 import org.camunda.bpm.engine.HistoryService;
 import org.camunda.bpm.engine.ManagementService;
-import org.camunda.bpm.engine.ProcessEngineConfiguration;
 import org.camunda.bpm.engine.RuntimeService;
-import org.camunda.bpm.engine.history.HistoricIncident;
 import org.camunda.bpm.engine.runtime.Job;
-import org.camunda.bpm.engine.runtime.JobQuery;
 import org.camunda.bpm.engine.runtime.ProcessInstance;
 import org.camunda.bpm.engine.test.Deployment;
 import org.camunda.bpm.engine.test.util.ProcessEngineBootstrapRule;
 import org.camunda.bpm.engine.test.util.ProcessEngineTestRule;
 import org.camunda.bpm.engine.test.util.ProvidedProcessEngineRule;
-import org.camunda.bpm.engine.test.util.Removable;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.ClassRule;
@@ -58,21 +53,17 @@ public class JobRetryCmdWithDefaultPropertyTest {
   protected RuntimeService runtimeService;
   protected ManagementService managementService;
   protected HistoryService historyService;
-  protected Removable removable;
 
   @Before
   public void setUp() {
     historyService = engineRule.getHistoryService();
     runtimeService = engineRule.getRuntimeService();
     managementService = engineRule.getManagementService();
-    removable = Removable.of(testRule);
   }
 
   @After
   public void tearDown() throws Exception {
     testRule.deleteHistoryCleanupJobs();
-    removable.remove(HistoricIncident.class);
-    resetHistoryCleanupConfig();
   }
 
   /**
@@ -109,45 +100,5 @@ public class JobRetryCmdWithDefaultPropertyTest {
 
     job = managementService.createJobQuery().jobId(job.getId()).singleResult();
     assertEquals(4, job.getRetries());
-  }
-
-  @Deployment(resources = { "org/camunda/bpm/engine/test/bpmn/async/FoxJobRetryCmdTest.testFailedServiceTask.bpmn20.xml" })
-  @Test
-  public void shouldApplyCleanupJobRetries() {
-    //given
-    ProcessInstance pi = runtimeService.startProcessInstanceByKey("failedServiceTask");
-
-    JobQuery jobQuery = managementService.createJobQuery().processInstanceId(pi.getProcessInstanceId());
-    Job job = jobQuery.singleResult();
-
-    assertNotNull(job);
-    assertEquals(2, job.getRetries()); // one try already failed
-
-    ProcessEngineConfiguration configuration = engineRule.getProcessEngineConfiguration();
-    //when
-    configuration.setHistoryCleanupDefaultNumberOfRetries(22);
-
-    //then
-    Job cleanupJob = historyService.cleanUpHistoryAsync(true);
-
-    assertThat(cleanupJob.getRetries()).isEqualTo(22);
-  }
-
-  @Deployment(resources = { "org/camunda/bpm/engine/test/bpmn/async/FoxJobRetryCmdTest.testFailedServiceTask.bpmn20.xml" })
-  @Test
-  public void shouldDisableRetriesOnCleanupJob() {
-    //given
-    ProcessEngineConfiguration configuration = engineRule.getProcessEngineConfiguration();
-    configuration.setHistoryCleanupDefaultNumberOfRetries(0);
-
-    //when
-    Job cleanupJob = historyService.cleanUpHistoryAsync(true);
-
-    //then
-    assertThat(cleanupJob.getRetries()).isEqualTo(0);
-  }
-
-  private void resetHistoryCleanupConfig() {
-    engineRule.getProcessEngineConfiguration().setHistoryCleanupDefaultNumberOfRetries(3);
   }
 }
